update DB and add upload file service

+++ src/main/java/fpt/se1601/giveandget/controller/MemberController.java
@@ -9,8 +9,6 @@ import fpt.se1601.giveandget.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
 public class MemberController {
     @Autowired
@@ -24,22 +22,31 @@ public class MemberController {
         DonationEntity donationEntity = donationService.addDonation(donationRequest);
         if (donationEntity == null)
             return "Add donation failed";
-        if (donationService.addRelationship(userService.findUserIdByToken(token), donationEntity.getId(),(short)1) == null)
+        if (donationService.addRelationship(userService.findUserIdByToken(token), donationEntity.getId(), (short) 1) == null)
             return "Add relationship failed but add donation successfully";
         return "Add donation successfully";
     }
 
     @PutMapping(value = "/donation")
-    public String updateDonation(@RequestBody DonationRequest donationRequest) {
-        if (donationService.updateDonation(donationRequest) != null)
-            return "Update donation success";
-        else
+    public String updateDonation(@RequestHeader("Authorization") String token, @RequestBody DonationRequest donationRequest) {
+        try {
+            if(!userService.isDonationOfUser(userService.findUserIdByToken(token),donationRequest.getId()))
+                return "It is not your donation";
+            if (donationService.updateDonation(donationRequest) != null)
+                return "Update donation success";
+            else
+                return "Update donation failed";
+        } catch (Exception e) {
+            e.printStackTrace();
             return "Update donation failed";
+        }
     }
 
     @DeleteMapping(value = "/donation")
-    public String deleteDonation(@RequestBody int id) {
+    public String deleteDonation(@RequestHeader("Authorization") String token,@RequestBody int id) {
         try {
+            if(!userService.isDonationOfUser(userService.findUserIdByToken(token),id))
+                return "It is not your donation";
             if (donationService.deleteDonation(id) != null)
                 return "Update donation success";
             return "Update donation failed";
@@ -48,12 +55,24 @@ public class MemberController {
             return "Update donation failed";
         }
     }
-    @PutMapping
-    public UserEntity updateUserInfo(UserInfoRequest userInfoRequest){
-        try{
-            return userService.updateUserInfo(userInfoRequest);
+
+    @GetMapping(value = "/user")
+    public UserEntity getUserInfo(@RequestHeader("Authorization") String token) {
+        try {
+            return userService.getUserByToken(token);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
         }
-        catch (Exception e){
+    }
+
+    @PutMapping(value = "/user")
+    public UserEntity updateUserInfo(@RequestHeader("Authorization") String token, @RequestBody UserInfoRequest userInfoRequest) {
+        try {
+            if(userService.findUserIdByToken(token) == userInfoRequest.getId())
+            return userService.updateUserInfo(userInfoRequest);
+            return null;
+        } catch (Exception e) {
             e.printStackTrace();
             return null;
         }
